changing the file structure to better fit in the file into eeitt-admin-frontend
